backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
email-validator==2.3.0
